requirements.txt
Flask
supabase
python-dotenv
gunicorn
openai
requests
pandas
openpyxl
tqdm